+++ lib/ui/views/home/home_viewmodel.dart
@@ -3,7 +3,7 @@ import 'package:stacked_services/stacked_services.dart';
 import 'package:veggie_go_malaysia/app/locator.dart';
 import 'package:veggie_go_malaysia/app/router.dart';
 import 'package:veggie_go_malaysia/datamodels/announcement.dart';
-import 'package:veggie_go_malaysia/services/database/abstract_firestore.dart';
+import 'package:veggie_go_malaysia/datamodels/place.dart';
 import 'package:veggie_go_malaysia/services/database/mock_firestore.dart';
 
 class HomeViewModel extends FutureViewModel {
@@ -13,6 +13,9 @@ class HomeViewModel extends FutureViewModel {
   List<Announcement> _announcements = <Announcement>[];
   List<Announcement> get announcements => _announcements;
 
+  List<Place> _nearestPlaces = <Place>[];
+  List<Place> get nearestPlaces => _nearestPlaces;
+
   void navigateToSearch() async {
     await _navigationService.navigateTo(Routes.searchViewRoute);
   }
@@ -20,5 +23,7 @@ class HomeViewModel extends FutureViewModel {
   @override
   Future futureToRun() async {
     _announcements = await _firestoreService.getAnnouncements();
+    _nearestPlaces = await _firestoreService.getNearestPlaces();
+    print(_nearestPlaces);
   }
 }
